python_backend/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.43.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==4.9.3
macholib==1.16.3
MarkupSafe==2.1.5
packaging==23.2
Pillow==10.0.1
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pytest==7.4.2
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzlocal==5.2
uvicorn==0.27.1
Werkzeug==3.0.1
